pyproject.toml
[project]
name = "airflow-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.6",
    "delta-spark>=3.1.0",
    "pyspark>=3.5.1",
]
